feat(async): add test to avoid panics on unexpected shutdown
This adds a test which ensures that unexpected shutdowns don't lead to panics.
Since we use the unhandled_panic function of Tokio which is currently part of
tokio_unstable, we have to enable this cfg when building the tests.
Sadly, .cargo/config.toml can't yet handle [target.'cfg(test)'] which would
make this more straight forward to use.