fix(engineio): respect pingTimeout and pingInterval to stop connection
The Engine.IO protocol features a heartbeat mechanism which ensures that the
connection between server and client is alive. During that heartbeat, the
server sends PINGs to which the client responds with PONGs. Both parties
can therefore detect whether the connection is still alive, based on the
pingInterval and pingTimeout received in the initial handshake.

However, we previously didn't make use of that in the Engine.IO implementation,
which lead to disconnects not being properly recognized.

We now respect these settings and return an error and disconnect the connection,
once the pingInterval+pingTimeout time has passed.

See also https://socket.io/docs/v4/how-it-works/#disconnection-detection